refactor: move `SelectEntityTypePage` to shared folder
