+++ apps/hash-frontend/src/pages/[shortname]/entities/[entity-uuid].page/select-entity-type-page.tsx
@@ -12,9 +12,9 @@ import { useContext, useState } from "react";
 
 import { useSnackbar } from "../../../../components/hooks/use-snackbar";
 import { Button } from "../../../../shared/ui";
+import { EntityTypeSelector } from "../../../shared/entity-type-selector";
 import { WorkspaceContext } from "../../../shared/workspace-context";
 import { SectionWrapper } from "../../shared/section-wrapper";
-import { EntityTypeSelector } from "./create-entity-page/entity-type-selector";
 import { EntityPageWrapper } from "./entity-page-wrapper";
 import { EntityPageHeader } from "./entity-page-wrapper/entity-page-header";
 import { LinksSectionEmptyState } from "./shared/links-section-empty-state";
